add package for file with index
